Use HTTP for external links
Previously, almost all external links were missing a scheme in their
URLs. Now that the website is hosted on github.io, the website can be
accessed with HTTPS. All links should have their scheme explicitly
specified as HTTP.

Some links may be obsolete as they are quite old, but no connectivities
are checked in this commit. A scheme is just added to all links.